refactor: move VxScan's test app screen to `libs/ui`
I'm keeping the old and new in place for now because VxMark still uses
the old one, but I'll update it soon.
